+++ src/packages/components/src/components/editors/DateTimePickerComponent.tsx
@@ -132,7 +132,7 @@ const DateTimePickerComponent: FC<DateTimePickerComponentProps> = ({
   }, [showDropdown])
 
   return (
-    <FieldContainer colspan={colspan} label={label} errorText={errorText}>
+    <FieldContainer colspan={colspan} label={label} errorText={errorText} disabled={disabled} readOnly={readOnly}>
       <div className={'relative'} ref={dropdownRef}>
         <input
           type={'text'}